requirements.txt
pexpect
unidecode
jsonrpclib
gensim
pypuzzle
google-api-python-client
python-twitter
pyvirtualdisplay
selenium